refactor: move functions to AuthService and utils
